Create Sum 1
Take a string and n number of rows as input . Arrange each character of string in a grid of n x l (l=length of string ) such that first character is first row first column ,second character in second row second column and on hitting the last row characters are inserted in rows in upward direction. Again on hitting the topmost row the characters are inserted in a downward direction and this repeats until all characters are entered.