Sum3 C++ code
Mogo and Pogo live in a city of grid. Mogo wants to meet Pogo ,so Pogo sends her the horizontal and vertical coordinates of his house to Pogo. You have to help Mogo find the number of steps in the shortest route from Mogo,s house to Pogo.s house . Assume Mogo stays at (0,0). Take input as number of testcases and coordinates of Pogo's house.